+++ CHANGELOG.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- None
+- Added [logi2dot] enable legend block.
+- Added [mdgrep] grep changelog example.
 
 
 ## [3.7.0] - 2023-04-07